+++ brewtils/test/fixtures.py
@@ -16,6 +16,7 @@ from brewtils.models import (
     Instance,
     IntervalTrigger,
     Job,
+    LegacyRole,
     LoggingConfig,
     Operation,
     Parameter,
@@ -26,7 +27,6 @@ from brewtils.models import (
     RequestFile,
     RequestTemplate,
     Resolvable,
-    LegacyRole,
     Runner,
     System,
 )
@@ -170,6 +170,7 @@ def command_dict(parameter_dict, system_id):
         "template": "<html></html>",
         "icon_name": "icon!",
         "metadata": {"meta": "data"},
+        "tags": [],
         "topics": [],
         "allow_any_kwargs": False,
     }
@@ -257,6 +258,7 @@ def system_dict(instance_dict, command_dict, command_dict_2, system_id):
         "namespace": "ns",
         "local": True,
         "template": "<html>template</html>",
+        "groups": ["GroupB", "GroupA"],
     }